backend/requirements.txt
fastapi
uvicorn
transformers
torch
pydantic
python-multipart
mangum
python-dotenv
sentencepiece
sacremoses
google-generativeai
openai
